Use associative array for matching
This change speeds up performance of matching stubs by using
an associative array for storing the matchers for a mock object.

Part of this change adds a getter method to IMethodMatcher,
which means it will force any class that implements this
interface to implement getMethod().  This makes this change a breaking
change, so it should be handled accordingly.

This change adde